requirements.txt
##############################
# Core Runtime & API Layer   #
##############################
fastapi==0.115.0
uvicorn==0.30.6

##############################
# Configuration / Validation #
##############################
pydantic==2.7.0
python-dotenv==1.0.1  # Load .env for config (optional but common)

##############################
# ML / Numeric Stack         #
##############################
tensorflow==2.20.0   # TF/Keras model loading & inference
numpy==1.26.4        # Array math
pandas==2.2.2        # CSV loading & preprocessing
joblib==1.4.2        # Optional: support .pkl models

##############################
# Dev / Prod Servers         #
##############################
# Use uvicorn to run FastAPI both locally and in prod

##############################
# Tooling / Testing (Optional)
##############################
pytest==8.2.0

##############################
# Optional Explainability   #
##############################
shap==0.45.1            # Model explainability (SHAP values)

rich==13.7.1            # Rich terminal output for scripts

##############################
# Installation Notes
##############################
# Create & activate venv, then install:
#   python -m venv .venv
#   . .venv/Scripts/Activate.ps1   (PowerShell on Windows)
#   python -m pip install --upgrade pip
#   pip install -r backend/requirements.txt

# If you do not need dotenv or a Windows WSGI server you can remove those lines.
